start to add mg and mgres solver

+++ codes/2dpe-shenliu/src/precision.f90
@@ -1,6 +1,6 @@
 module precision
-    use, intrinsic :: iso_fortran_env, &
-        si => int16, di => int32, fi => int8, &
+    use, intrinsic :: iso_fortran_env, only : &
+        si => int32, di => int64, fi => int16, &
         sp => real32, dp => real64
     implicit none
 end module precision
